+++ src/services/api/getArticleList/index.ts
@@ -1,10 +1,14 @@
 import { getHttpClient } from "@/helpers/httpClient";
 
-const BASE_URL = 'api/queryDatabase';
+const BASE_URL = 'queryDatabase';
 
 export async function getArticleList() {
   const httpClient = await getHttpClient();
   const resp = await httpClient.post(BASE_URL);
-
-  return resp.data;
+  
+  try{
+    return await resp.data;
+  }catch(err) {
+    return Promise.reject(err)
+  }
 }