+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## CHANGELOG
 
+## 24/04/2022
+- [bug] Gọi api bị cancle lý do config sai domain, không export default function tại folder api
 ## 08/04/2022
 - [core] update env for application && add structure for api